requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-logging
google-cloud-monitoring

